requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.4
django-cleanup==6.0.0
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
multidict==6.0.3
Pillow==9.3.0
Pygments==2.13.0
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sqlparse==0.4.3
urllib3==1.26.13
